backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
django_environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2024.1
psycopg2-binary==2.9.9
psycopg==3.1.18
python-decouple==3.8
google_api_python_client==2.120.0
pillow == 10.3.0









